Fixed some bugs / Improved profiler

+++ include/simd/simd-table.h
@@ -42,8 +42,8 @@
 #define	M_PI6	(M_PI4*M_PI2)
 /* Only single precision for the moment */
 
-constexpr size_t iSgAd = 0b1000000000000000000000000000000000000000000000000000000000000000;
-constexpr size_t iSgAb = 0b1000000000000000000000000000000010000000000000000000000000000000;
+constexpr long long int iSgAd = 0b1000000000000000000000000000000000000000000000000000000000000000;
+constexpr long long int iSgAb = 0b1000000000000000000000000000000010000000000000000000000000000000;
 #ifdef	__AVX512F__
 constexpr _MInt_  iSgnAbsd  = {       iSgAd,       iSgAd,       iSgAd,       iSgAd,       iSgAd,       iSgAd,       iSgAd,       iSgAd };
 constexpr _MInt_  iSgnAbsf  = {       iSgAb,       iSgAb,       iSgAb,       iSgAb,       iSgAb,       iSgAb,       iSgAb,       iSgAb };
@@ -60,9 +60,9 @@ constexpr _MInt_  iSgnAbsf  = {       iSgAb,       iSgAb };
 constexpr double Inf_d = __builtin_inf();
 constexpr double Nan_d = __builtin_nan("");//0xFFFFF");
 
-constexpr size_t m32Mk = 0b0000000000000000000000000001111100000000000000000000000000011111;
-constexpr size_t iSgMk = 0b1100000000000000000000000000000011000000000000000000000000000000;
-constexpr size_t iFlMk = 0b1111100000000000000000000000000011111000000000000000000000000000;
+constexpr long long int m32Mk = 0b0000000000000000000000000001111100000000000000000000000000011111;
+constexpr long long int iSgMk = 0b1100000000000000000000000000000011000000000000000000000000000000;
+constexpr long long int iFlMk = 0b1111100000000000000000000000000011111000000000000000000000000000;
 
 /*	For the exponential	*/
 constexpr double th1_d  = +1.80911414126145723458e+03;
